--- packages/react/src/components/index.js
@@ -1,5 +1,0 @@
-export { ChatBody } from './ChatBody';
-export { ChatHeader } from './ChatHeader';
-export { ChatInput } from './ChatInput';
-export { EmojiPicker } from './EmojiPicker';
-export { Home } from './Home';
