+++ packages/react/src/store/attachmentwindow.js
@@ -1,4 +1,4 @@
-import create from 'zustand';
+import { create } from 'zustand';
 
 const useAttachmentWindowStore = create((set) => ({
   open: false,
