+++ packages/react/src/store/loginmodalStore.js
@@ -1,4 +1,4 @@
-import create from 'zustand';
+import { create } from 'zustand';
 
 const loginModalStore = create((set) => ({
   isLoginModalOpen: false,
