+++ packages/react/src/store/mentionmemberStore.js
@@ -1,4 +1,4 @@
-import create from 'zustand';
+import { create } from 'zustand';
 
 const mentionmemberStore = create((set) => ({
   roomMembers: {},
