+++ packages/react/src/store/searchMessageStore.js
@@ -1,4 +1,4 @@
-import create from 'zustand';
+import { create } from 'zustand';
 
 const useSearchMessageStore = create((set) => ({
   showSearch: false,
