+++ packages/react/src/store/totpmodalStore.js
@@ -1,4 +1,4 @@
-import create from 'zustand';
+import { create } from 'zustand';
 
 const totpModalStore = create((set) => ({
   isModalOpen: false,
